Refactor instance generator into modular files
Split generator.rs into config, policies, err, and mod.rs for clarity
and extensibility. Add support for multiple/disjoint space windows in
config. Requests now merge overlapping/adjacent space windows on
creation.
